memchr: reorganize crate
This reorganizes the internal structure of the crate to be less of a
mess, including rewriting the tests. In particular, we commit to no_std
and obsolete the need for the `use_std` feature, since we simply elect
to always use `core`.

We also add an explicit compilation error for non-32/64 bit systems. We
technically supported it previously, but it's not clear if it ever
worked since it was never tested. We can add it back if necessary.
